Add an input_filter for TextEdit.

+++ crates/egui/src/widgets/text_edit/builder.rs
@@ -58,6 +58,7 @@ pub struct TextEdit<'t> {
     font_selection: FontSelection,
     text_color: Option<Color32>,
     layouter: Option<&'t mut dyn FnMut(&Ui, &str, f32) -> Arc<Galley>>,
+    input_filter: Option<&'t mut dyn FnMut(&str) -> bool>,
     password: bool,
     frame: bool,
     margin: Vec2,
@@ -103,6 +104,7 @@ impl<'t> TextEdit<'t> {
             font_selection: Default::default(),
             text_color: None,
             layouter: None,
+            input_filter: None,
             password: false,
             frame: true,
             margin: vec2(4.0, 2.0),
@@ -213,6 +215,12 @@ impl<'t> TextEdit<'t> {
     /// ```
     pub fn layouter(mut self, layouter: &'t mut dyn FnMut(&Ui, &str, f32) -> Arc<Galley>) -> Self {
         self.layouter = Some(layouter);
+
+        self
+    }
+
+    pub fn input_filter(mut self, input_filter: &'t mut dyn FnMut(&str) -> bool) -> Self {
+        self.input_filter = Some(input_filter);
 
         self
     }
@@ -362,6 +370,7 @@ impl<'t> TextEdit<'t> {
             font_selection,
             text_color,
             layouter,
+            input_filter,
             password,
             frame: _,
             margin: _,
@@ -404,6 +413,10 @@ impl<'t> TextEdit<'t> {
         let layouter = layouter.unwrap_or(&mut default_layouter);
 
         let mut galley = layouter(ui, text.as_str(), wrap_width);
+        let mut default_input_filter = |_: &str| {
+            true
+        };
+        let input_filter = input_filter.unwrap_or(&mut default_input_filter);
 
         let desired_width = if multiline {
             galley.size().x.max(wrap_width) // always show everything in multiline
@@ -532,6 +545,7 @@ impl<'t> TextEdit<'t> {
                 text,
                 &mut galley,
                 layouter,
+                input_filter,
                 id,
                 wrap_width,
                 multiline,
@@ -691,6 +705,7 @@ fn events(
     text: &mut dyn TextBuffer,
     galley: &mut Arc<Galley>,
     layouter: &mut dyn FnMut(&Ui, &str, f32) -> Arc<Galley>,
+    input_filter: &mut dyn FnMut(&str) -> bool,
     id: Id,
     wrap_width: f32,
     multiline: bool,
@@ -735,7 +750,7 @@ fn events(
                 }
             }
             Event::Paste(text_to_insert) => {
-                if !text_to_insert.is_empty() {
+                if !text_to_insert.is_empty() && input_filter(text_to_insert) {
                     let mut ccursor = delete_selected(text, &cursor_range);
                     insert_text(&mut ccursor, text, text_to_insert);
                     Some(CCursorRange::one(ccursor))
@@ -745,7 +760,7 @@ fn events(
             }
             Event::Text(text_to_insert) => {
                 // Newlines are handled by `Key::Enter`.
-                if !text_to_insert.is_empty() && text_to_insert != "\n" && text_to_insert != "\r" {
+                if !text_to_insert.is_empty() && text_to_insert != "\n" && text_to_insert != "\r" && input_filter(text_to_insert) {
                     let mut ccursor = delete_selected(text, &cursor_range);
                     insert_text(&mut ccursor, text, text_to_insert);
                     Some(CCursorRange::one(ccursor))
@@ -758,7 +773,7 @@ fn events(
                 pressed: true,
                 modifiers,
             } => {
-                if multiline && ui.memory().has_lock_focus(id) {
+                if multiline && ui.memory().has_lock_focus(id) && input_filter("\t") {
                     let mut ccursor = delete_selected(text, &cursor_range);
                     if modifiers.shift {
                         // TODO(emilk): support removing indentation over a selection?
@@ -816,7 +831,7 @@ fn events(
             }
 
             Event::CompositionUpdate(text_mark) => {
-                if !text_mark.is_empty() && text_mark != "\n" && text_mark != "\r" && state.has_ime
+                if !text_mark.is_empty() && text_mark != "\n" && text_mark != "\r" && state.has_ime && input_filter(text_mark)
                 {
                     let mut ccursor = delete_selected(text, &cursor_range);
                     let start_cursor = ccursor;
@@ -832,6 +847,7 @@ fn events(
                     && prediction != "\n"
                     && prediction != "\r"
                     && state.has_ime
+                    && input_filter(prediction)
                 {
                     state.has_ime = false;
                     let mut ccursor = delete_selected(text, &cursor_range);
